Feature Request Template
Adds a Feature Request Template for when contributors or anyone else has an idea or multiple ideas or proposals/multiple proposals, so they can plan out the ideas or proposals with the SLR team. Basically, request features, or you can request the feature and someone from the SLR can give you an answer to your idea or proposal.